+++ sql-planner/src/builtins.lua
@@ -3,6 +3,52 @@ local dt = require('datetime')
 -- Builtin sbroad funcs implemented in LUA
 local builtins = {}
 
+builtins._PICO_BUCKET = function(tier)
+    local op = "_pico_bucket"
+
+    if tier == nil then
+        return error(("%s: tier name is null"):format(op))
+    end
+
+    local info = _G.pico.router[tier]
+    if info == nil then
+        return error(("%s: tier '%s' doesn't exist"):format(op, tier))
+    end
+
+    local max_bucket_id = info:bucket_count()
+    local result = {}
+
+    for bucket_id = 1, max_bucket_id do
+        local rs, err = info:route(bucket_id)
+        if rs == nil then
+            return error(("%s: %s"):format(op, err.message))
+        end
+
+        local last = result[#result]
+        if last
+            and rs.uuid == last.current_replicaset_uuid
+            and bucket_id == last.bucket_id_end + 1
+        then
+            -- extend the existing range
+            last.bucket_id_end = bucket_id
+        else
+            -- create a new entry and record it as the last for this uuid
+            local entry = {
+                tier_name               = tier,
+                bucket_id_start         = bucket_id,
+                bucket_id_end           = bucket_id,
+                state                   = 'active',
+                current_replicaset_uuid = rs.uuid,
+                target_replicaset_uuid  = rs.uuid,
+            }
+
+            result[#result + 1] = entry
+        end
+    end
+
+    return result
+end
+
 builtins.TO_DATE = function(s, fmt)
     if s == nil or fmt == nil then
         return nil
@@ -51,6 +97,18 @@ local function init()
     end
 
     _G[module].builtins = builtins
+
+    body = string.format("function(...) return %s.builtins._PICO_BUCKET(...) end",
+        module)
+    box.schema.func.create("_pico_bucket", {
+        language = 'LUA',
+        returns = 'ARRAY',
+        body = body,
+        param_list = { 'string' },
+        exports = { 'SQL' },
+        is_deterministic = false,
+        if_not_exists = true
+    })
 
     -- We don't want to make builtin functions global,
     -- so we put them in module. But for function